+++ internal/adapters/dataproviders/database/mongo/pegout.go
@@ -328,8 +328,8 @@ func (repo *pegoutMongoRepository) ListQuotesByDateRange(ctx context.Context, st
 
 	dbCtx, cancel := context.WithTimeout(ctx, repo.conn.timeout)
 	defer cancel()
-	startTimestamp := uint32(startDate.Unix())
-	endTimestamp := uint32(endDate.Unix())
+	startTimestamp := int64(startDate.Unix())
+	endTimestamp := int64(endDate.Unix())
 	quoteCollection := repo.conn.Collection(PegoutQuoteCollection)
 	quoteFilter := bson.D{
 		primitive.E{
